added refactoring
