+++ homework_30/core/__init__.py
@@ -1,3 +1,4 @@
 from .base_model import Base
 from .config import Config
 from .singleton import Singleton
+from .infrastracture import BaseRepository
