+++ homework_30/core/infrastracture/__init__.py
@@ -1,0 +1,1 @@
+from .base_repository import BaseRepository
